requirements.txt
streamlit
yt-dlp
openai-whisper
torch
transformers
docx
python-docx
